Added: get profile complete functional
